backend_pirates_way_finder/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
networkx
pymongo
pydantic[email]
python-jose
passlib[argon2]
python-multipart
Pillow
matplotlib
numpy